- update readme and create english version - refactor part of the code into 3 files
